Updated json-adapter tests
