+++ test/scripts/adapters/index.js
@@ -1,0 +1,5 @@
+// jshint esversion: 8
+
+describe('Adapters', () => {
+  require('./json');
+});
