+++ test/scripts/adapters/json/async/index.js
@@ -1,0 +1,9 @@
+// jshint esversion: 8
+
+// Test createDatabaseValue
+describe('JsonFileAdapter', () => {
+
+  require('./save');
+  require('./load');
+
+});
