+++ test/scripts/adapters/json/index.js
@@ -1,0 +1,6 @@
+// jshint esversion: 8
+
+describe('json', () => {
+  require('./async');
+  require('./sync');
+});
