+++ test/scripts/adapters/json/sync/index.js
@@ -1,0 +1,9 @@
+// jshint esversion: 8
+
+// Test createDatabaseValue
+describe('SyncJsonFileAdapter', () => {
+
+  require('./save');
+  require('./load');
+
+});
